Refactor out MerchantDetail from MerchantSideBar
