+++ web/src/components/merchant/ongoing-listings/index.tsx
@@ -17,9 +17,9 @@
 import React from 'react';
 
 import MerchantSideBar from 'components/common/MerchantSideBar';
-import styled from 'styled-components';
 import EmptyListingsPlaceholder from 'components/merchant/ongoing-listings/EmptyListingsPlaceholder';
 import Row from 'muicss/lib/react/row';
+import styled from 'styled-components';
 
 const PageContainer = styled.div`
   min-height: 100vh;
